Speed up checking whether an image is already saved
Under the assumption that an image name without extension is unique,
there's no need to determine the file type to check if the file is
already saved. A check for the basename plus any extension suffices.
This saves a whole roundtrip for images without extension.

Experiments show that the file's basename changes even when the
picture in a photo post is edited, so this should be safe.