tumblr_backup: provide command-line help
This means to replace getopt with the more modern (but already deprecated)
optparse. The reason why I don't use the current argparse is that it was
only added to Python's standard library in release 2.7. Anyway, optparse
is relatively close to argparse API-wise, so most of the work can be reused.