tumblr_backup: sort the posts by ID before checking limits
This fixes the problem that back-dated posts were not saved by the
incremental backup mode. When a post is back-dated, its ID is higher
than the currently most recent post (IDs always increase), but it
is not the first one in the XML (the XML is ordered by descending time).

Sorting the posts by ID descending solves this.