tumblr_backup: unify media download
All media are saved in media/. Use the same method to download photos,
audio and Tumblr-hosted video. Media are downloaded in 1MB-sized chunks.

NOTE: The old 'images' directory is left as-is, because old saved posts
may still point to images saved there. If you intend to do a new
complete backup, it makes sense to rename 'images' to 'media' first, so
that the images already downloaded can be reused.