tumblr_backup: simplify -Q syntax
WARNING: This commit breaks existing scripts that use -Q.

Change the TAG separator to colon, and the rule separator to comma.
The new rule syntax is now TYPE:TAG:TAG:…,TYPE:TAG:…,….

This is a win because of two factors:

  - The semicolons between the rules had to be quoted in typical Unix
    shells.

  - Using the same separator inside one rule makes parsing the rule
    easier.

In effect, the -T and -Q options are now equivalent. But don't tell
anyone, because the manual doesn't say so ;-)

Also add documentation for -t and -Q.