Special instructions on youtube-dl install/maintence
I've submitted multiple issues for youtube-dl that were found through (but not in) tumblr_backup.py. Embeds for vid.me, Pornhub, and two for Tumblr it's self. All regex errors. If users have youtube-dl installed, they need to update it and be good about updating it so the script doesn't hang or miss content.

So this is a detailed instruction on how to install it and keep to up to date. I tried to keep the instruction as short as possible, only elaborating on the problem of switching from the script you get from the website to the one from pip. Figure it saves some the legwork of trying to find the solution when their script won't update.

Theres a lot of bug fixes coming that will stop tumblr_backup.py from failing on blogs with video on them.