tumblr_backup: Fix excessive memory usage while building the index
Posts are now read from disk each time they are needed, instead of
keeping them all in memory.

This is especially needed when --save-notes is used with a large blog,
as this often results in many gigabytes worth of HTML that does not fit
into free memory all at once.