tumblr_backup: Remove thread killing code
I apparently overlooked some information in [1], which says:

  However, if the target thread is executing the Python interpreter,
  the Python signal handlers will be executed by the main thread of the
  main interpreter. Therefore, the only point of sending a signal to a
  particular Python thread would be to force a running system call to
  fail with InterruptedError.

Further, [2] notes:

  Changed in version 3.5: Python now retries system calls when a syscall
  is interrupted by a signal, except if the signal handler raises an
  exception (see PEP 475 for the rationale), instead of raising
  InterruptedError.

So after Python 3.5, there is no use for signal.pthread_kill. It cannot
do what it does in C, which is send a signal to a spefic thread - at
best, the thread will wake up with EINTR and then resume the system
call.

Also, the sys.exc_info() check was wrong: There will always be an active
KeyboardInterrupt exception in a signal handler. This means that the
pthread_kill() never had any effect, masking the fact that only the main
thread received the signal.

The signal handlers are being kept because they are still useful if the
program receives SIGTERM or SIGHUP, allowing resources to be cleaned up
via exception handling instead of being left for the OS to deal with.

[1] https://docs.python.org/3/library/signal.html#signal.pthread_kill
[2] https://docs.python.org/3/library/exceptions.html#InterruptedError